Starting implimentations of REST see diff files for changes. TODO:  test that XML is correctly created  <result> 1 </result or   <result> 0</result> the XML also needs to be parsed form $output in the login-exe.php file, the parsed XML should be stored in $result  where it is tested like normal.  if result == '1' then a session will be created. else login failed
made a chance to password formt he post... since not https post is plain text :(  minimal security i md5 hash the password before sending it to the authenticator. this will take some load off the function since it will be comparing to an already hashed password on the sql database.